353142: Separated the Writer and Read Role Scripts so that permissions can be assigned using the correct roles (#126)
In order to grant the database permissions correctly, two container app jobs are required. One to grant permissions for the Writer role and then another to grant permissions for the Reader role (using the Writer role). The reason for this is that the Writer role is the owner of the database tables and it is this role that can grant read permissions to other roles to read those tables.

The changes in this PR are to split the permissions into two separate scripts.